Modify size_t to int about blocksize

+++ src/mode/ECB.h
@@ -3,4 +3,4 @@
 #include <stdint.h>
 #include <string.h>
 
-void ECB(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *key, size_t block_size, size_t len, unsigned char *src, unsigned char *des);
+void ECB(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *key, int block_size, size_t len, unsigned char *src, unsigned char *des);
